rename `logWriterClass.ts` -> `logWriter.ts`
