+++ src/logWriter.ts
